Introduce a new constructfrom generic function (#28)
* Introduce a new constructfrom generic function

An alternative, more generic approach to constructing generic objects
from other objects that also respects defined `StructTypes.StructType`
properties of both source and target objects. The usage is pretty
simple, just `StructTypes.constructfrom(T, obj)`, so give the type `T`
you want to construct, and a compatible object `obj`. `constructfrom`
will call `StructType` on the target type `T` and then proceed to
construct it appropriately, whether building up an array, dict, or
constructing a struct or mutable struct.

This is an alternative approach to the `makeobj` approach that existed
briefly. `constructfrom` is more generic in that it supports all
`StructType` types, and utilizes the `foreachfield`, `construct`, and
`applyfield!` utility functions to ensure serialization/deserialization
names are respected, keyword args, etc.