[BREAKING] Add new StructTypes.OrderedStruct and StructTypes.UnorderedStruct (#41)
Both of these are subtypes of StructTypes.Struct. To ease transition for
existing users, `StructTypes.Struct()` will default to
`StructTypes.UnorderedStruct`. StructTypes dependents will need to
update to support this new major version release, but _users_ of the
package shouldn't need to do anything but bump their version as existing
behavior of `StructTypes.Struct` will remain valid/the same. We also
change the StructType of `NamedTuple`s to `StructTypes.Struct`, which
resolves #39. We also fix and add tests for singletons reading/writing
which resolves #9.