Pass `nothing` for missing struct fields in `construct` (#43)
* Pass `nothing` for missing struct fields in `construct`

This primarily affects `StructTypes.UnorderedStruct` routines that end
up calling `StructTypes.construct` with fields passed in a
`Vector{Any}`. Before `UnorderedStruct` was introduced,
`StructTypes.Struct` structs were passed whatever arguments found when
deserializing in order to the constructor. With the introduction of
`UnorderedStruct`, we now aim to allow implementations to pass fields in
the correct order, regardless of how they were serialized. This poses a
problem, however, for fields not present in the serialized data, as is
often the case for JSON. It's typical to "omit null" fields when
serializing, but then when deserializing, we had the awkward scenario of
how to treat those not-present fields for `UnorderedStruct`
constructors. The proposal here is to pass `nothing` for not-present
fields, which mirrors the common pattern of "omitting null"s when
serializing, we're just providing the JSON equivalent of `null`,
`nothing` for that field. This should simplify `StructTypes.Struct`
custom type constructor's lives since they can now know to expect the
same number of arguments any time their type is deserialized; some may
be `nothing` if they are not present when deserializing. Fixes
https://github.com/quinnj/JSON3.jl/issues/136.

* docs